refacto: split html util
