+++ src/util/mod.rs
@@ -3,17 +3,18 @@ pub mod condition;
 mod context;
 pub mod fonts;
 mod header;
-mod html;
 mod size;
 mod spacing;
+mod style;
+mod tag;
 
 pub use attributes::{suffix_css_classes, suffix_unit, Attributes};
 pub use context::Context;
 pub use header::Header;
-pub use html::Style;
-pub use html::Tag;
 pub use size::Size;
 pub use spacing::Spacing;
+pub use style::Style;
+pub use tag::Tag;
 
 use std::cmp::Ordering;
